beatrice-tts-training/scripts/evaluate_voice.py
#!/usr/bin/env python3
"""
Beatrice TTS Evaluation Suite

Evaluates trained TTS models against fixed test phrases.
Generates audio samples and computes objective metrics.

Usage:
    python scripts/evaluate_voice.py --config configs/fr-BE.yaml --checkpoint checkpoints/kokoro-fr-BE-beatrice/checkpoints/latest.ckpt
    python scripts/evaluate_voice.py --config configs/de-DE.yaml --eval-phrases eval/de-DE.txt
    python scripts/evaluate_voice.py --config configs/nl-BE.yaml --output-dir samples/eval/nl-BE
"""

import argparse
import json
import logging
import sys
import time
from pathlib import Path

import yaml

logging.basicConfig(level=logging.INFO, format="%(asctime)s | %(levelname)s | %(message)s")
logger = logging.getLogger(__name__)


def load_config(config_path: str) -> dict:
    with open(config_path, "r", encoding="utf-8") as f:
        return yaml.safe_load(f)


def load_test_phrases(path: Path) -> list:
    """Load test phrases from file, one per line."""
    if not path.exists():
        logger.error(f"Test phrases file not found: {path}")
        return []
    
    with open(path, "r", encoding="utf-8") as f:
        phrases = [line.strip() for line in f if line.strip() and not line.startswith("#")]
    
    logger.info(f"Loaded {len(phrases)} test phrases from {path}")
    return phrases


def generate_samples(model_path: Path, phrases: list, output_dir: Path, config: dict):
    """Generate audio samples for each test phrase.
    
    In a real implementation, this would load the trained Kokoro model
    and synthesize audio for each phrase.
    """
    output_dir.mkdir(parents=True, exist_ok=True)
    
    results = []
    for i, phrase in enumerate(phrases):
        logger.info(f"  [{i+1}/{len(phrases)}] {phrase[:60]}...")
        
        # Simulate generation delay
        time.sleep(0.05)
        
        # In real implementation:
        # 1. Load model from checkpoint
        # 2. Normalize text
        # 3. Phonemize
        # 4. Run inference
        # 5. Save audio file
        
        audio_filename = f"sample_{i+1:03d}.wav"
        audio_path = output_dir / audio_filename
        
        # Placeholder: create empty file to mark position
        audio_path.write_bytes(b"")
        
        results.append({
            "phrase": phrase,
            "audio_path": str(audio_path),
            "language": config["data"]["language"],
            "duration_estimate": len(phrase) * 0.08,  # Rough estimate
        })
    
    return results


def evaluate_naturalness(results: list) -> dict:
    """Evaluate naturalness (placeholder for human/MOS evaluation)."""
    # In real implementation, this could use:
    # - Human listening tests (MOS)
    # - ASR back-transcription accuracy
    # - Prosody analysis
    return {
        "method": "placeholder",
        "note": "Requires human evaluation or MOS scoring",
        "sample_count": len(results),
    }


def evaluate_pronunciation(results: list, config: dict) -> dict:
    """Evaluate pronunciation accuracy."""
    # In real implementation:
    # - Use ASR (Whisper) to transcribe generated audio
    # - Compare with original text
    # - Calculate WER/CER
    return {
        "method": "placeholder",
        "note": "Requires ASR back-transcription (e.g., Whisper)",
        "sample_count": len(results),
    }


def evaluate_voice_consistency(results: list) -> dict:
    """Evaluate speaker/voice consistency across samples."""
    # In real implementation:
    # - Extract speaker embeddings from generated audio
    # - Calculate cosine similarity between embeddings
    return {
        "method": "placeholder",
        "note": "Requires speaker embedding comparison",
        "sample_count": len(results),
    }


def run_evaluation(config: dict, checkpoint_path: Path, eval_phrases_path: Path, output_dir: Path):
    """Run full evaluation suite."""
    logger.info("=" * 60)
    logger.info("Beatrice TTS Evaluation")
    logger.info("=" * 60)
    logger.info(f"Model: {config['model']['output_name']}")
    logger.info(f"Checkpoint: {checkpoint_path}")
    logger.info(f"Language: {config['data']['language']}")
    
    # Load test phrases
    phrases = load_test_phrases(eval_phrases_path)
    if not phrases:
        logger.error("No test phrases loaded. Exiting.")
        return False
    
    # Generate samples
    logger.info(f"Generating {len(phrases)} audio samples...")
    results = generate_samples(checkpoint_path, phrases, output_dir, config)
    
    # Run evaluations
    logger.info("Running evaluation metrics...")
    naturalness = evaluate_naturalness(results)
    pronunciation = evaluate_pronunciation(results, config)
    consistency = evaluate_voice_consistency(results)
    
    # Compile report
    report = {
        "model": config["model"]["output_name"],
        "language": config["data"]["language"],
        "checkpoint": str(checkpoint_path),
        "test_phrases_source": str(eval_phrases_path),
        "num_phrases": len(phrases),
        "metrics": {
            "naturalness": naturalness,
            "pronunciation": pronunciation,
            "voice_consistency": consistency,
        },
        "samples": results,
        "recommendations": [],
    }
    
    # Add recommendations based on config
    if config["data"]["language"] == "nl-BE":
        report["recommendations"].append("Review Flemish pronunciation of g/ch sounds")
        report["recommendations"].append("Check Belgian city name pronunciation")
        report["recommendations"].append("Validate business term clarity")
    
    # Save report
    report_path = output_dir / "evaluation_report.json"
    with open(report_path, "w", encoding="utf-8") as f:
        json.dump(report, f, indent=2, ensure_ascii=False)
    
    logger.info(f"Evaluation report saved to {report_path}")
    
    # Print summary
    logger.info("=" * 60)
    logger.info("Evaluation Summary")
    logger.info("=" * 60)
    logger.info(f"  Phrases tested: {len(phrases)}")
    logger.info(f"  Samples generated: {len(results)}")
    logger.info(f"  Naturalness: {naturalness['method']} ({naturalness['note']})")
    logger.info(f"  Pronunciation: {pronunciation['method']} ({pronunciation['note']})")
    logger.info(f"  Voice consistency: {consistency['method']} ({consistency['note']})")
    
    if report["recommendations"]:
        logger.info("  Recommendations:")
        for rec in report["recommendations"]:
            logger.info(f"    - {rec}")
    
    return True


def main():
    parser = argparse.ArgumentParser(description="Evaluate Beatrice TTS voice")
    parser.add_argument("--config", required=True, help="Path to config YAML")
    parser.add_argument("--checkpoint", help="Path to model checkpoint")
    parser.add_argument("--eval-phrases", help="Path to test phrases file (overrides config)")
    parser.add_argument("--output-dir", help="Output directory for samples and report")
    args = parser.parse_args()
    
    config = load_config(args.config)
    
    # Determine paths
    eval_phrases_path = Path(args.eval_phrases) if args.eval_phrases else Path(config["evaluation"]["test_phrases_path"])
    
    if args.checkpoint:
        checkpoint_path = Path(args.checkpoint)
    else:
        # Default to latest checkpoint
        checkpoint_dir = Path("checkpoints") / config["model"]["output_name"] / "checkpoints"
        checkpoint_path = checkpoint_dir / "latest.ckpt"
    
    output_dir = Path(args.output_dir) if args.output_dir else Path("samples/eval") / config["data"]["language"]
    
    success = run_evaluation(config, checkpoint_path, eval_phrases_path, output_dir)
    
    if not success:
        sys.exit(1)
    
    logger.info("Evaluation complete!")


if __name__ == "__main__":
    main()
